+++ ui/shared/components/panel/family/FamilyFields.jsx
@@ -1,10 +1,10 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import { connect } from 'react-redux'
-import { Popup } from 'semantic-ui-react'
+import { Popup, Icon } from 'semantic-ui-react'
 import styled from 'styled-components'
 
-import { loadUserOptions, loadProjectAnalysisGroups } from 'redux/rootReducer'
+import { loadUserOptions, loadProjectAnalysisGroups, updateFamily } from 'redux/rootReducer'
 import {
   getSamplesByFamily,
   getUserOptionsIsLoading,
@@ -14,7 +14,9 @@ import {
   getAnalysisGroupsByFamily,
   getAnalysisGroupIsLoading,
 } from 'redux/selectors'
+import { SNP_DATA_TYPE, FAMILY_ANALYSED_BY_DATA_TYPES } from 'shared/utils/constants'
 
+import DispatchRequestButton from '../../buttons/DispatchRequestButton'
 import TagFieldView from '../view-fields/TagFieldView'
 import Sample from '../sample'
 import { ColoredIcon } from '../../StyledComponents'
@@ -108,40 +110,48 @@ export const analysisStatusIcon = (
   )
 }
 
-const formatAnalysedByList = analysedByList => analysedByList.map(
-  analysedBy => `${analysedBy.createdBy.displayName || analysedBy.createdBy.email} (${new Date(analysedBy.lastModifiedDate).toLocaleDateString()})`,
-).join(', ')
+const BaseAnalysedBy = React.memo(({ analysedByList, compact, onSubmit }) => {
+  const analysedByType = analysedByList.reduce(
+    (acc, analysedBy) => ({ ...acc, [analysedBy.dataType]: [...(acc[analysedBy.dataType] || []), analysedBy] }), {},
+  )
 
-export const AnalysedBy = React.memo(({ analysedByList, compact }) => {
   if (compact) {
-    return [...analysedByList.reduce(
-      (acc, analysedBy) => acc.add(analysedBy.createdBy.displayName || analysedBy.createdBy.email), new Set(),
+    return [...(analysedByType[SNP_DATA_TYPE] || []).reduce(
+      (acc, { createdBy }) => acc.add(createdBy), new Set(),
     )].map(
       analysedByUser => <NoWrap key={analysedByUser}>{analysedByUser}</NoWrap>,
     )
   }
-  const analystUsers = analysedByList.filter(analysedBy => analysedBy.createdBy.isAnalyst)
-  const externalUsers = analysedByList.filter(analysedBy => !analysedBy.createdBy.isAnalyst)
-  return [
-    analystUsers.length > 0 ? (
-      <div key="analyst">
-        <b>CMG Analysts:</b>
-        {formatAnalysedByList(analystUsers)}
-      </div>
-    ) : null,
-    externalUsers.length > 0 ? (
-      <div key="ext">
-        <b>External Collaborators:</b>
-        {formatAnalysedByList(externalUsers)}
-      </div>
-    ) : null,
-  ]
+
+  return FAMILY_ANALYSED_BY_DATA_TYPES.map(([type, typeDisplay]) => (
+    <div key={type}>
+      <b>{`${typeDisplay}: `}</b>
+      {(analysedByType[type] || []).map(
+        analysedBy => `${analysedBy.createdBy} (${new Date(analysedBy.lastModifiedDate).toLocaleDateString()})`,
+      ).join(', ')}
+      &nbsp;&nbsp;
+      <DispatchRequestButton
+        buttonContent={<Icon link size="small" name="plus" />}
+        onSubmit={onSubmit(type)}
+        confirmDialog={`Are you sure you want to add that you analysed this family for ${typeDisplay} data?`}
+      />
+    </div>
+  ))
 })
 
-AnalysedBy.propTypes = {
+BaseAnalysedBy.propTypes = {
   analysedByList: PropTypes.arrayOf(PropTypes.object),
   compact: PropTypes.bool,
+  onSubmit: PropTypes.func,
 }
+
+const mapDispatchToProps = (dispatch, ownProps) => ({
+  onSubmit: dataType => () => dispatch(
+    updateFamily({ dataType, familyGuid: ownProps.familyGuid, familyField: 'analysed_by' }),
+  ),
+})
+
+export const AnalysedBy = connect(null, mapDispatchToProps)(BaseAnalysedBy)
 
 const BaseAnalysisGroups = React.memo(({ load, loading, ...props }) => (
   <DataLoader load={load} loading={loading} content>